Explictly include jest-environment-jsdom
Starting from Jest 28, jest-environment-jsdom is no longer included by default in the Jest package. It needs to be installed separately when you want to test code that interacts with the DOM (like React components). Since we're looking at React and WordPress block editor components, we may need a DOM environment for testing.